+++ docs/source/developer_guides/custom_models.md
@@ -104,7 +104,7 @@ Finally, we can use any training framework we like, or write our own fit loop, t
 
 For a complete example, check out [this notebook](https://github.com/huggingface/peft/blob/main/examples/multilayer_perceptron/multilayer_perceptron_lora.ipynb).
 
-## timm model
+## timm models
 
 The [timm](https://huggingface.co/docs/timm/index) library contains a large number of pretrained computer vision models.
 Those can also be fine-tuned with PEFT. Let's check out how this works in practice.
@@ -199,3 +199,13 @@ peft_model.print_trainable_parameters()
 This shows us that we only need to train less than 2% of all parameters, which is a huge efficiency gain.
 
 For a complete example, check out [this notebook](https://github.com/huggingface/peft/blob/main/examples/image_classification/image_classification_timm_peft_lora.ipynb).
+
+## New transformers architectures
+
+When new popular transformers architectures are released, we will do our best to quickly add them to PEFT. Should you still come across a transformers model that is not supported out of the box, a good idea is to check the existing models for inspiration. You can find them inside of [constants.py](https://github.com/huggingface/peft/blob/main/src/peft/utils/constants.py) in the PEFT repository. Often, you'll find a similar architecture that uses the same names.
+
+If that doesn't help, check the existing modules as described above and try to identify the attention layers, especially the key, query, and value layers. Those will often have names such as `c_attn`, `query`, `q_proj`, etc. Note that the "key" layer is not always adapted, ideally you can check for yourself if including it results in better performance.
+
+Additionally, linear layers are common targets to be adapted. Their names will often contain the strings `fc` or `dense`.
+
+If you want to add a new model to PEFT, please create an entry in [constants.py](https://github.com/huggingface/peft/blob/main/src/peft/utils/constants.py) and open a pull request on the [repository](https://github.com/huggingface/peft/pulls). Don't forget to update the [README](https://github.com/huggingface/peft#models-support-matrix) as well.
